dijkstra is now working with euclidian distances as weights

+++ src/algorithms/Dijkstra.h
@@ -18,7 +18,7 @@ namespace vOS
         OpenVolumeMesh::VertexHandle m_start;
         OpenVolumeMesh::VertexHandle m_end;
         OpenVolumeMesh::GeometricPolyhedralMeshV3f m_mesh;
-        OpenVolumeMesh::EdgePropertyT<int> m_weights;
+        OpenVolumeMesh::EdgePropertyT<float> m_weights;
 
         void PauseButtonReleased();
         void PauseButtonPressed();
